// types/agents.ts
export interface Agents {
    _id: string;
    agentId: string;
    agentName: string;
    phone: string;
    email: string;
    region: string;
    target: Number;
    achieved: Number;
    status: string;
    createdAt: string;
    imageUrl: string;
    from: Date;
    to: Date;
    conversion: string;
}